Added LMS Mentor Service

+++ src/main/java/com/bridgelabz/lmsmentorservice/service/MentorService.java
@@ -27,7 +27,7 @@ public class MentorService implements IMentorService{
 
     @Override
    public Response addMentor(MentorDTO mentorDTO, String token) {
-        boolean isUserPresent = restTemplate.getForObject("http://localhost:8080/admin/validate/" + token, Boolean.class);
+        boolean isUserPresent = restTemplate.getForObject("http://localhost:8082/admin/validate/" + token, Boolean.class);
         if (isUserPresent) {
             MentorModel mentorModel = new MentorModel(mentorDTO);
             Long mentorId = tokenUtil.decodeToken(token);
@@ -36,7 +36,7 @@ public class MentorService implements IMentorService{
             String body = "Mentor added Successfully with id  :" + mentorModel.getId();
             String subject = "Mentor added Successfully....";
             mailService.send(mentorModel.getEmail(), body, subject);
-            return new Response("Mentor Added Successfully", 200, mentorId);
+            return new Response("Mentor Added Successfully", 200, mentorModel);
         }
         throw new AdminException(400, "Token is Wrong");
     }
@@ -66,40 +66,46 @@ public class MentorService implements IMentorService{
 
     @Override
    public Response updateMentor(long id, MentorDTO mentorDTO, String token) {
-        Long mentorId = tokenUtil.decodeToken(token);
-        Optional<MentorModel> isMentorPresent = mentorRepository.findById(mentorId);
-        if (isMentorPresent.isPresent()){
-            isMentorPresent.get().setEmployeeId(mentorDTO.getEmployeeId());
-            isMentorPresent.get().setFirstName(mentorDTO.getFirstName());
-            isMentorPresent.get().setLastName(mentorDTO.getLastName());
-            isMentorPresent.get().setMentorType(mentorDTO.getMentorType());
-            isMentorPresent.get().setMobileNum(mentorDTO.getMobileNum());
-            isMentorPresent.get().setEmail(mentorDTO.getEmail());
-            isMentorPresent.get().setExperienceYear(mentorDTO.getExperienceYear());
-            isMentorPresent.get().setPreferredTime(mentorDTO.getPreferredTime());
-            isMentorPresent.get().setStartDate(mentorDTO.getStartDate());
-            isMentorPresent.get().setStatus(mentorDTO.getStatus());
-            isMentorPresent.get().setMentorDescription(mentorDTO.getMentorDescription());
-            isMentorPresent.get().setProfileImageUrl(mentorDTO.getProfileImageUrl());
-            isMentorPresent.get().setCreatorUser(mentorDTO.getCreatorUser());
-            isMentorPresent.get().setSupervisor(mentorDTO.getSupervisor());
-            isMentorPresent.get().setCreatedTimeStamp(mentorDTO.getCreatedTimeStamp());
-            isMentorPresent.get().setUpdatedTimeStamp(mentorDTO.getUpdatedTimeStamp());
-            String body = "Mentor updated Successfully with id  :" + isMentorPresent.get().getId();
-            String subject = "Mentor updated Successfully....";
-            mailService.send(isMentorPresent.get().getEmail(), body, subject);
-            return new Response("Mentor Added Successfully", 200, isMentorPresent);
+        boolean isUserPresent = restTemplate.getForObject("http://localhost:8080/admin/validate/" + token, Boolean.class);
+        if (isUserPresent) {
+            Long mentorId = tokenUtil.decodeToken(token);
+            Optional<MentorModel> isMentorPresent = mentorRepository.findById(mentorId);
+            if (isMentorPresent.isPresent()) {
+                isMentorPresent.get().setEmployeeId(mentorDTO.getEmployeeId());
+                isMentorPresent.get().setFirstName(mentorDTO.getFirstName());
+                isMentorPresent.get().setLastName(mentorDTO.getLastName());
+                isMentorPresent.get().setMentorType(mentorDTO.getMentorType());
+                isMentorPresent.get().setMobileNum(mentorDTO.getMobileNum());
+                isMentorPresent.get().setEmail(mentorDTO.getEmail());
+                isMentorPresent.get().setExperienceYear(mentorDTO.getExperienceYear());
+                isMentorPresent.get().setPreferredTime(mentorDTO.getPreferredTime());
+                isMentorPresent.get().setStartDate(mentorDTO.getStartDate());
+                isMentorPresent.get().setStatus(mentorDTO.getStatus());
+                isMentorPresent.get().setMentorDescription(mentorDTO.getMentorDescription());
+                isMentorPresent.get().setProfileImageUrl(mentorDTO.getProfileImageUrl());
+                isMentorPresent.get().setCreatorUser(mentorDTO.getCreatorUser());
+                isMentorPresent.get().setSupervisor(mentorDTO.getSupervisor());
+                isMentorPresent.get().setCreatedTimeStamp(mentorDTO.getCreatedTimeStamp());
+                isMentorPresent.get().setUpdatedTimeStamp(mentorDTO.getUpdatedTimeStamp());
+                String body = "Mentor updated Successfully with id  :" + isMentorPresent.get().getId();
+                String subject = "Mentor updated Successfully....";
+                mailService.send(isMentorPresent.get().getEmail(), body, subject);
+                return new Response("Mentor Added Successfully", 200, isMentorPresent);
+            }
         }
         throw  new AdminException(400, "Mentor not Present");
     }
 
     @Override
    public Response deleteMentor(Long id, String token) {
-        Long mentorId = tokenUtil.decodeToken(token);
-        Optional<MentorModel> isMentorPresent = mentorRepository.findById(mentorId);
-        if (isMentorPresent.isPresent()){
-            mentorRepository.delete(isMentorPresent.get());
-            return new Response("Mentor Added Successfully", 200, isMentorPresent);
+        boolean isUserPresent = restTemplate.getForObject("http://localhost:8080/admin/validate/" + token, Boolean.class);
+        if (isUserPresent) {
+            Long mentorId = tokenUtil.decodeToken(token);
+            Optional<MentorModel> isMentorPresent = mentorRepository.findById(mentorId);
+            if (isMentorPresent.isPresent()) {
+                mentorRepository.delete(isMentorPresent.get());
+                return new Response("Mentor Added Successfully", 200, isMentorPresent);
+            }
         }
         throw new AdminException(400, "Mentor Not found");
     }
